Move to reading the readme directly from the file
Since PyPI now accepts markdown and not just restructured text we can
just pass the file. This also solves a problem where pandoc was
struggling to convert some html into rst.

As a bonus we remove a build-time dependency on pandoc.

We also remove the changelog, but this is visible in the github releases
page anyway.