[bugfix] Use TextDecoder to decode resources using UTF-8
The standard `fromCharCode` we were using to decode
character-by-character was using UTF-16 by default. This was a problem
when we had utf-8 encoded resources and so we use TextDecoder to decode
these.

However, the TextDecoder API is not available in wkhtmltox, which is the
reason for disabling the compression entirely when using wkhtmltox.