Use browsers' DecompressionStream to do zlib decompression instead of rolling our own
We had our own zlib decompression library in jsinflate taken from
http://www.onicos.com/staff/iz/amuse/javascript/expert/inflate.txt
originlly (now at
https://www.masanao.site/staff/iz/amuse/javascript/expert/inflate.txt).

However using the DecompressionStream API available in modern browsers
we can completely remove this dependency that is injected into _every_
block (unless we explicitly opt out of compression). This decreases
block sizes dramatically for small blocks, and opens us up to
performance improvements with native implementations of the zlib
algorithm.

It is also the case that in many situations the decompression code is
far longer than the script we are decompressing and so this attempt at
minimising the size of a block in bytes actually makes it larger! Now,
with a far shorter decompression script, this is less likely and we may
actually see benefits from compression (though b64 encoding may again
make it larger than just shipping the bytes raw).