Add PNG regression tests
Our current regression suite checks that the rendered HTML doesn't
change. However, sometimes we want to change the HTML (and so will bless
the output), but don't expect the rendered _image_ to change. This test
makes sure that we are aware of any pixel changes in the image output of
pybloqs.

Note: This makes the regression testing now _very_ sensitive,
particularly to third party changes. We will need to freeze all the
versions of our dependencies that we use in order for this to not become
unnecessarily noisy.