[server] Permit specifying the endpoint that blocks serve on
By passing construction parameters to the BloqsProvider through the
decorator bloqs_provider, we can set the _id variable and thus the
endpoint that the server serves each bloq on.

We make the decorator accept optional parameters so that you can call it
"bare" or with parameters, keeping backwards compatibility with previous
bare-only usage.