Replace entire div with pybloqs provider response
When we render a provider without calling it, it places a small div in
the document that loads the actual contents from `self.url` when it is
revealed. Previously we were replacing the innerHTML (default) which
left `hx-get` as a property on a parent div, which could then be
inherited by its children.

What we should be doing is replacing the entire small div with the
contents of the block.